adjust migrations

+++ database/migrations/2025_09_22_072101_create_discounts_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('discounts', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->decimal('percentage');
+            $table->integer('percentage');
             $table->timestamps();
         });
     }
